chore: address feedback
